Turn viewer and viewport service into core services.
Credit goes to @aaron on whose PR this is based. I realized, though, that viewer and viewport would be pretty hard and frustrating, so I chose to finish it myself.

Part of #1055
